+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_help.py
@@ -89,7 +89,8 @@ helps['acs kubernetes'] = """
 
 helps['acs kubernetes get-credentials'] = """
     type: command
-    short-summary: Download and install credentials to access a cluster.
+    short-summary: Download and install credentials to access a cluster.  This command requires
+                   the same private-key used to create the cluster.
 """
 
 helps['acs list-locations'] = """
@@ -122,12 +123,12 @@ helps['acs wait'] = """
 
 helps['aks'] = """
      type: group
-     short-summary: Manage Azure Kubernetes Services.
+     short-summary: (PREVIEW) Manage Azure Kubernetes Services.
 """
 
 helps['aks browse'] = """
     type: command
-    short-summary: Show the dashboard for a Kubernetes cluster in a web browser.
+    short-summary: (PREVIEW) Show the dashboard for a Kubernetes cluster in a web browser.
     parameters:
         - name: --disable-browser
           type: bool
@@ -137,7 +138,7 @@ helps['aks browse'] = """
 
 helps['aks create'] = """
     type: command
-    short-summary: Create a new managed Kubernetes cluster.
+    short-summary: (PREVIEW) Create a new managed Kubernetes cluster.
     parameters:
         - name: --generate-ssh-keys
           type: string
@@ -188,12 +189,12 @@ helps['aks create'] = """
 
 helps['aks delete'] = """
     type: command
-    short-summary: Delete a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Delete a managed Kubernetes cluster.
 """
 
 helps['aks get-credentials'] = """
     type: command
-    short-summary: Get access credentials for a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Get access credentials for a managed Kubernetes cluster.
     parameters:
         - name: --admin -a
           type: bool
@@ -205,22 +206,22 @@ helps['aks get-credentials'] = """
 
 helps['aks get-upgrades'] = """
     type: command
-    short-summary: Get the upgrade versions available for a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Get the upgrade versions available for a managed Kubernetes cluster.
 """
 
 helps['aks get-versions'] = """
     type: command
-    short-summary: Get the versions available for creating a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Get the versions available for creating a managed Kubernetes cluster.
 """
 
 helps['aks install-cli'] = """
     type: command
-    short-summary: Download and install kubectl, the Kubernetes command-line tool.
+    short-summary: (PREVIEW) Download and install kubectl, the Kubernetes command-line tool.
 """
 
 helps['aks install-connector'] = """
     type: command
-    short-summary: Install the ACI Connector on a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Install the ACI Connector on a managed Kubernetes cluster.
     parameters:
         - name: --chart-url
           type: string
@@ -277,12 +278,12 @@ helps['aks install-connector'] = """
 
 helps['aks list'] = """
     type: command
-    short-summary: List managed Kubernetes clusters.
+    short-summary: (PREVIEW) List managed Kubernetes clusters.
 """
 
 helps['aks remove-connector'] = """
     type: command
-    short-summary: Remove the ACI Connector from a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Remove the ACI Connector from a managed Kubernetes cluster.
     parameters:
         - name: --connector-name
           type: string
@@ -302,7 +303,7 @@ helps['aks remove-connector'] = """
 
 helps['aks scale'] = """
     type: command
-    short-summary: Scale the node pool in a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Scale the node pool in a managed Kubernetes cluster.
     parameters:
         - name: --node-count -c
           type: int
@@ -311,12 +312,12 @@ helps['aks scale'] = """
 
 helps['aks show'] = """
     type: command
-    short-summary: Show the details for a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Show the details for a managed Kubernetes cluster.
 """
 
 helps['aks upgrade'] = """
     type: command
-    short-summary: Upgrade a managed Kubernetes cluster to a newer version.
+    short-summary: (PREVIEW) Upgrade a managed Kubernetes cluster to a newer version.
     long-summary: "Kubernetes will be unavailable during cluster upgrades."
     parameters:
         - name: --kubernetes-version -k
@@ -328,7 +329,7 @@ helps['aks upgrade'] = """
 
 helps['aks upgrade-connector'] = """
     type: command
-    short-summary: Upgrade the ACI Connector on a managed Kubernetes cluster.
+    short-summary: (PREVIEW) Upgrade the ACI Connector on a managed Kubernetes cluster.
     parameters:
         - name: --chart-url
           type: string
@@ -385,7 +386,7 @@ helps['aks upgrade-connector'] = """
 
 helps['aks wait'] = """
     type: command
-    short-summary: Wait for a managed Kubernetes cluster to reach a desired state.
+    short-summary: (PREVIEW) Wait for a managed Kubernetes cluster to reach a desired state.
     long-summary: If an operation on a cluster was interrupted or was started with `--no-wait`, use this command to
                   wait for it to complete.
     examples: