+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_help.py
@@ -72,16 +72,6 @@ helps['functionapp identity assign'] = helps['webapp identity assign'].replace('
 
 helps['functionapp identity show'] = helps['webapp identity show'].replace('webapp', 'functionapp')
 
-helps['webapp assign-identity'] = """
-    type: command
-    short-summary: (Deprecated, please use 'az webapp identity assign')
-"""
-
-helps['functionapp assign-identity'] = """
-    type: command
-    short-summary: (Deprecated, please use 'az functionapp identity assign')
-"""
-
 helps['webapp config'] = """
 type: group
 short-summary: Configure a web app.