+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/custom.py
@@ -26,18 +26,20 @@ except ImportError:
     pass
 import websocket
 from knack.log import get_logger
-from knack.prompting import prompt_pass, NoTTYException
+from knack.prompting import prompt_pass, prompt, NoTTYException
 from knack.util import CLIError
 from azure.mgmt.containerinstance.models import (AzureFileVolume, Container, ContainerGroup, ContainerGroupNetworkProtocol,
                                                  ContainerPort, ImageRegistryCredential, IpAddress, Port, ResourceRequests,
-                                                 ResourceRequirements, Volume, VolumeMount, ContainerExecRequestTerminalSize)
+                                                 ResourceRequirements, Volume, VolumeMount, ContainerExecRequestTerminalSize,
+                                                 GitRepoVolume)
 from ._client_factory import cf_container_groups, cf_container_logs, cf_start_container
 
 logger = get_logger(__name__)
 WINDOWS_NAME = 'Windows'
-ACR_SERVER_SUFFIX = '.azurecr.io/'
+SERVER_DELIMITER = '.'
 AZURE_FILE_VOLUME_NAME = 'azurefile'
 SECRETS_VOLUME_NAME = 'secrets'
+GITREPO_VOLUME_NAME = 'gitrepo'
 
 
 def list_containers(client, resource_group_name=None):
@@ -78,6 +80,10 @@ def create_container(client,
                      azure_file_volume_account_name=None,
                      azure_file_volume_account_key=None,
                      azure_file_volume_mount_path=None,
+                     gitrepo_url=None,
+                     gitrepo_dir='.',
+                     gitrepo_revision=None,
+                     gitrepo_mount_path=None,
                      secrets=None,
                      secrets_mount_path=None):
     """Create a container group. """
@@ -113,6 +119,13 @@ def create_container(client,
     if secrets_volume:
         volumes.append(secrets_volume)
         mounts.append(secrets_volume_mount)
+
+    gitrepo_volume = _create_gitrepo_volume(gitrepo_url=gitrepo_url, gitrepo_dir=gitrepo_dir, gitrepo_revision=gitrepo_revision)
+    gitrepo_volume_mount = _create_gitrepo_volume_mount(gitrepo_volume=gitrepo_volume, gitrepo_mount_path=gitrepo_mount_path)
+
+    if gitrepo_volume:
+        volumes.append(gitrepo_volume)
+        mounts.append(gitrepo_volume_mount)
 
     cgroup_ip_address = _create_ip_address(ip_address, ports, dns_name_label)
 
@@ -158,7 +171,13 @@ def _create_image_registry_credentials(registry_login_server, registry_username,
         image_registry_credentials = [ImageRegistryCredential(server=registry_login_server,
                                                               username=registry_username,
                                                               password=registry_password)]
-    elif ACR_SERVER_SUFFIX in image:
+    elif SERVER_DELIMITER in image.split("/")[0]:
+        if not registry_username:
+            try:
+                registry_username = prompt(msg='Image registry username: ')
+            except NoTTYException:
+                raise CLIError('Please specify --registry-username in order to use Azure Container Registry.')
+
         if not registry_password:
             try:
                 registry_password = prompt_pass(msg='Image registry password: ')
@@ -166,13 +185,12 @@ def _create_image_registry_credentials(registry_login_server, registry_username,
                 raise CLIError('Please specify --registry-password in order to use Azure Container Registry.')
 
         acr_server = image.split("/")[0] if image.split("/") else None
-        acr_username = image.split(ACR_SERVER_SUFFIX)[0] if image.split(ACR_SERVER_SUFFIX) else None
-        if acr_server and acr_username:
+        if acr_server:
             image_registry_credentials = [ImageRegistryCredential(server=acr_server,
-                                                                  username=acr_username,
+                                                                  username=registry_username,
                                                                   password=registry_password)]
         else:
-            raise CLIError('Failed to parse ACR server or username from image name; please explicitly specify --registry-server and --registry-username.')
+            raise CLIError('Failed to parse ACR server from image name; please explicitly specify --registry-server.')
 
     return image_registry_credentials
 
@@ -201,6 +219,13 @@ def _create_secrets_volume(secrets):
     return Volume(name=SECRETS_VOLUME_NAME, secret=secrets) if secrets else None
 
 
+def _create_gitrepo_volume(gitrepo_url, gitrepo_dir, gitrepo_revision):
+    """Create Git Repo volume. """
+    gitrepo_volume = GitRepoVolume(repository=gitrepo_url, directory=gitrepo_dir, revision=gitrepo_revision)
+
+    return Volume(name=GITREPO_VOLUME_NAME, git_repo=gitrepo_volume) if gitrepo_url else None
+
+
 # pylint: disable=inconsistent-return-statements
 def _create_azure_file_volume_mount(azure_file_volume, azure_file_volume_mount_path):
     """Create Azure File volume mount. """
@@ -218,6 +243,15 @@ def _create_secrets_volume_mount(secrets_volume, secrets_mount_path):
             raise CLIError('Please specify --secrets --secrets-mount-path '
                            'to enable secrets volume mount.')
         return VolumeMount(name=SECRETS_VOLUME_NAME, mount_path=secrets_mount_path)
+
+
+def _create_gitrepo_volume_mount(gitrepo_volume, gitrepo_mount_path):
+    """Create Git Repo volume mount. """
+    if gitrepo_mount_path:
+        if not gitrepo_volume:
+            raise CLIError('Please specify --gitrepo-url (--gitrepo-dir --gitrepo-revision) '
+                           'to enable Git Repo volume mount.')
+        return VolumeMount(name=GITREPO_VOLUME_NAME, mount_path=gitrepo_mount_path)
 
 
 # pylint: disable=inconsistent-return-statements
@@ -250,14 +284,14 @@ def container_logs(cmd, resource_group_name, name, container_name=None, follow=F
             stream_args=(logs_client, resource_group_name, name, container_name, container_group.restart_policy))
 
 
-def container_exec(cmd, resource_group_name, name, container_name=None, exec_command=None, terminal_row_size=20, terminal_col_size=80):
+def container_exec(cmd, resource_group_name, name, exec_command, container_name=None, terminal_row_size=20, terminal_col_size=80):
     """Start exec for a container. """
 
     start_container_client = cf_start_container(cmd.cli_ctx)
     container_group_client = cf_container_groups(cmd.cli_ctx)
     container_group = container_group_client.get(resource_group_name, name)
 
-    if container_name is None and len(container_group.containers) == 1:
+    if container_name or container_name is None and len(container_group.containers) == 1:
         # If only one container in container group, use that container.
         if container_name is None:
             container_name = container_group.containers[0].name