+++ src/command_modules/azure-cli-billing/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.1.8
+++++++
+* Add enrollment account commands
+* `sdist` is now compatible with wheel 0.31.0
+
 0.1.7
 ++++++
 * Update for CLI core changes.