+++ src/command_modules/azure-cli-profile/HISTORY.rst
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+2.0.22
+++++++
+* account list: handle accounts which come from CLI 1.0/ASM mode
+* (Breaking change): remove `--msi` & `--msi-port` which were tagged `deprecating` 2 releases ago
+* `sdist` is now compatible with wheel 0.31.0
+
 2.0.21
 ++++++
 * az login: warn on using --identity-port/--msi-port as they become useless with imds support