+++ src/command_modules/azure-cli-rdbms/HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+0.2.2
++++++
+* Minor help fix.
+
+0.2.1
++++++
+* Introduce georestore command.
+* Remove storage size restriction from create command.
+* `sdist` is now compatible with wheel 0.31.0
+
 0.2.0
 +++++
 * Release with new business model GA API version 2017-12-01.